dry convertFromJsonCanvas, add clear btn
