+++ src/components/GenerateSpace.vue
@@ -7,7 +7,6 @@ import MinimapCanvas from '@/components/MinimapCanvas.vue'
 
 import dayjs from 'dayjs'
 import utils from '@/utils.js'
-import importUtils from '@/importUtils.js'
 import consts from '@/consts.js'
 import { nanoid } from 'nanoid'
 import randomColor from 'randomcolor'
@@ -21,7 +20,8 @@ const state = reactive({
   newSpace: null,
   size: null,
   pageWidth: null,
-  pageHeight: null
+  pageHeight: null,
+  errors: []
 })
 
 const isOnline = computed(() => store.state.isOnline)
@@ -148,17 +148,15 @@ const sample = {
   ]
 }
 
-// from import.vue
-// DRY to util or somthing
-
 const isValidCanvas = (space) => {
+  state.errors = []
   const schema = {
     'nodes': 'array',
     'edges': 'array'
   }
   validateSchema(space, schema)
   if (state.errors.length) {
-    throw new Error('🚒 isValid')
+    throw new Error('🚒 isValid', space)
   }
 }
 const validateSchema = (space, schema) => {
@@ -171,97 +169,6 @@ const validateSchema = (space, schema) => {
     }
   })
 }
-const convertFromCanvas = (space) => {
-  const minPositionValue = 150
-  let date = dayjs(new Date())
-  date = date.format(consts.nameDateFormat)
-  let newSpace = {}
-  try {
-    newSpace.name = `Canvas ${date}`
-    newSpace.id = nanoid()
-    newSpace.background = consts.defaultSpaceBackground
-    newSpace.cards = []
-    newSpace.connections = []
-    newSpace.connectionTypes = []
-    // emsure node positions are positive 0,0
-    let negativePositionOffset = {
-      x: 0,
-      y: 0
-    }
-    space.nodes.forEach(node => {
-      if (node.x < negativePositionOffset.x) {
-        negativePositionOffset.x = node.x
-      }
-      if (node.y < negativePositionOffset.y) {
-        negativePositionOffset.y = node.y
-      }
-    })
-    space.nodes = space.nodes.map(node => {
-      node.x = node.x + Math.abs(negativePositionOffset.x)
-      node.y = node.y + Math.abs(negativePositionOffset.y)
-      return node
-    })
-    // nodes → cards
-    const shouldNudgeCardsY = Boolean(space.nodes.find(node => node.y <= minPositionValue))
-    const shouldNudgeCardsX = Boolean(space.nodes.find(node => node.x <= minPositionValue))
-    space.nodes.forEach(node => {
-      // url
-      let shouldUpdateUrlPreview
-      if (node.url) {
-        shouldUpdateUrlPreview = true
-      }
-      // y
-      let y = node.y
-      if (shouldNudgeCardsY) {
-        y += minPositionValue
-      }
-      // x
-      let x = node.x
-      if (shouldNudgeCardsX) {
-        x += minPositionValue
-      }
-      // name
-      let name = node.text || node.url || node.label
-      if (node.file) {
-        name = `\`${node.file}\``
-      }
-      const newCard = {
-        id: node.id,
-        x,
-        y,
-        backgroundColor: node.canvasColor || node.color,
-        name,
-        shouldUpdateUrlPreview
-      }
-      newSpace.cards.push(newCard)
-    })
-
-    // connection type
-    const typeId = nanoid()
-    const newConnetionType = {
-      id: typeId,
-      color: newTypeColor(),
-      name: `Connection Type 0`
-    }
-    newSpace.connectionTypes.push(newConnetionType)
-    // edges → connections
-    space.edges.forEach((edge, index) => {
-      const newConnection = {
-        id: edge.id,
-        startItemId: edge.fromNode,
-        endItemId: edge.toNode,
-        controlPoint: `q00,00`, // straight line
-        directionIsVisible: Boolean(edge.fromEnd === 'arrow' || edge.toEnd === 'arrow'),
-        connectionTypeId: typeId,
-        labelIsVisible: Boolean(edge.label)
-      }
-      newSpace.connections.push(newConnection)
-    })
-    return newSpace
-  } catch (error) {
-    console.error('🚒 convertFromCanvas', error)
-  }
-}
 
 const newTypeColor = () => {
   const isThemeDark = store.state.currentUser.theme === 'dark'
@@ -271,8 +178,6 @@ const newTypeColor = () => {
   }
   return color
 }
-
-// -------
 
 const normalizeSpace = async (space) => {
   const { cards, boxes, connectionTypes, connections } = space
@@ -297,13 +202,15 @@ const generatePreview = async () => {
   if (state.isGeneratingPreview) { return }
   try {
     state.isGeneratingPreview = true
-    // call api w prompt and return canvasJson and spaceName
-    let space = convertFromCanvas(sample)
+    // TODO await call api w prompt and return canvasJson and spaceName
+    isValidCanvas(sample)
+    const typeColor = newTypeColor()
+    let space = utils.convertFromJsonCanvas(sample, typeColor)
     space = await normalizeSpace(space)
     space.name = 'NYC Move with Kids: Pros & Cons Map' // TODO
     updateSize(space)
     state.newSpace = space
-    console.log('🔮🔮🔮🔮🔮 newspace', state.newSpace)
+    console.log('🔮 generatePreview', space)
   } catch (error) {
     console.error('🚒 generatePreview', error)
   }
@@ -311,13 +218,16 @@ const generatePreview = async () => {
 }
 
 const minimapCanvasIsVisible = computed(() => Boolean(state.newSpace))
-
+const clear = () => {
+  state.newSpace = null
+}
 </script>
 
 <template lang="pug">
 section.generate-space(v-if="isOnline")
   input(placeholder="Type to generate a space")
   button(:class="{active: state.isGeneratingPreview}" @click="generatePreview")
+    img.icon.openai(src="@/assets/openai.svg")
     span Preview
     Loader(:visible="state.isGeneratingPreview")
   .minimap-canvas-inline-wrap(ref="rowElement")
@@ -329,9 +239,12 @@ section.generate-space(v-if="isOnline")
       :pageWidth="state.pageWidth"
       :viewportIsHidden="true"
       )
-    button.add-space-button(v-if="minimapCanvasIsVisible")
-      img.icon.add(src="@/assets/add.svg")
-      span {{state.newSpace.name}}
+    template(v-if="minimapCanvasIsVisible")
+      button.cancel-minimap.small-button(title="Cancel" @click="clear")
+        img.icon.cancel(src="@/assets/add.svg")
+      button.add-space-button
+        img.icon.add(src="@/assets/add.svg")
+        span {{state.newSpace.name}}
 </template>
 
 <style lang="stylus">
@@ -343,11 +256,10 @@ section.generate-space(v-if="isOnline")
   .add-space-button
     border-top-left-radius 0
     border-top-right-radius 0
-
-  // .minimap-canvas-inline-wrap
-  //   button
-  //     position absolute
-  //     bottom -16px
-  //     left 14px
-
+  .minimap-canvas-inline-wrap
+    position relative
+    .cancel-minimap
+      position absolute
+      top 6px
+      right 6px
 </style>
