--- src/importUtils.js
@@ -1,5 +1,0 @@
-import utils from '@/utils.js'
-
-export default {
-
-}
